refactor(postcss-merge-rules): split huge files
